test(annotate): replace round 34 acceptance with the 34b click-time-prep flow
The old spec encoded the superseded ready-gated dialog (disabled Start
button, gear reopen path). The new one covers the 34b contract: no gear,
tool-button click opens the dialog, Start annotating is enabled
immediately and spins only during click-time preparation, the box tool
responds right after the dialog closes, B activates directly once ready,
Recompute shows only on the selected model row when its embedding
exists, no nested-button DOM warnings, and an anonymous session gets a
surfaced sign-in error instead of an infinite spinner. baseURL is
overridable via DEV_BASE_URL.
